+++ server/hazard-service/src/main/java/edu/illinois/ncsa/incore/service/hazard/controllers/TornadoController.java
@@ -182,7 +182,6 @@ public class TornadoController {
     @Operation(summary = "Creates a new tornado, the newly created tornado is returned.",
         description = "Additionally, a GeoTiff (raster) is created by default and publish to data repository. " +
             "User can create both model tornadoes and dataset-based tornadoes with GeoTiff files uploaded.")
-
     @RequestBody(description = "Tornado json and files.", required = true,
         content = @Content(mediaType = MediaType.APPLICATION_FORM_URLENCODED,
             schema = @Schema(type = "object",
@@ -263,7 +262,14 @@ public class TornadoController {
 
                 // Store the dataset
                 datasetId = ServiceUtil.createDataset(datasetObject, this.username, this.userGroups, files);
-                tornadoModel.setDatasetId(datasetId);
+                // Assuming only one hazardDataset will be created here.
+                // construct tornado hazardDatasets
+                TornadoHazardDataset hazardDataset = new TornadoHazardDataset();
+                hazardDataset.setDatasetId(datasetId);
+                hazardDataset.setDemandType(TornadoHazard.DEMAND_TYPE);
+                hazardDataset.setDemandUnits(TornadoHazard.WIND_MPH);
+                hazardDataset.setThreshold(null);
+                tornadoModel.addTornadoHazardDataset(hazardDataset);
 
                 tornado.setCreator(this.username);
                 tornado.setOwner(this.username);
@@ -271,12 +277,13 @@ public class TornadoController {
                 addTornadoToSpace(tornado, this.username);
             } else if (tornado != null && tornado instanceof TornadoDataset) {
                 TornadoDataset tornadoDataset = (TornadoDataset) tornado;
-                datasetId = null;
+
                 if (fileParts != null && !fileParts.isEmpty() && TornadoUtils.validateDatasetTypes(fileParts)) {
                     // Create dataset object representation for storing shapefile
                     JSONObject datasetObject = TornadoUtils.getTornadoDatasetObject("Tornado Hazard", "EF Boxes representing tornado");
                     // Store the dataset
                     datasetId = ServiceUtil.createDataset(datasetObject, this.username, this.userGroups);
+
                     if (datasetId != null) {
                         // attach files to the dataset
                         int statusCode = ServiceUtil.attachFileToTornadoDataset(datasetId, this.username, this.userGroups, fileParts);
@@ -285,11 +292,18 @@ public class TornadoController {
                             logger.error(tornadoErrorMsg);
                             throw new IncoreHTTPException(Response.Status.BAD_REQUEST, tornadoErrorMsg);
                         }
+
+                        // construct tornado hazardDatasets
+                        TornadoHazardDataset hazardDataset = new TornadoHazardDataset();
+                        hazardDataset.setDatasetId(datasetId);
+                        hazardDataset.setDemandType(TornadoHazard.DEMAND_TYPE);
+                        hazardDataset.setDemandUnits(TornadoHazard.WIND_MPH);
+                        hazardDataset.setThreshold(null);
+                        tornadoDataset.addTornadoHazardDataset(hazardDataset);
                     } else {
                         logger.error(tornadoJsonErrorMsg);
                         throw new IncoreHTTPException(Response.Status.BAD_REQUEST, tornadoJsonErrorMsg);
                     }
-                    ((TornadoDataset) tornado).setDatasetId(datasetId);
 
                     tornado.setCreator(this.username);
                     tornado.setOwner(this.username);
@@ -314,7 +328,6 @@ public class TornadoController {
         }
         ServiceUtil.deleteDataset(datasetId, this.username, this.userGroups);
         throw new IncoreHTTPException(Response.Status.BAD_REQUEST, tornadoErrorMsg);
-
     }
 
     @GET
@@ -531,14 +544,17 @@ public class TornadoController {
             // delete associated datasets
             if (tornado != null && tornado instanceof TornadoModel) {
                 TornadoModel tModel = (TornadoModel) tornado;
-                if (ServiceUtil.deleteDataset(tModel.getDatasetId(), this.username, this.userGroups) == null) {
-                    spaceRepository.addToOrphansSpace(tModel.getDatasetId());
+                for (TornadoHazardDataset dataset: tModel.getHazardDatasets()) {
+                    if (ServiceUtil.deleteDataset(dataset.getDatasetId(), this.username, this.userGroups) == null) {
+                        spaceRepository.addToOrphansSpace(dataset.getDatasetId());
+                    }
                 }
             } else if (tornado != null && tornado instanceof TornadoDataset) {
                 TornadoDataset tDataset = (TornadoDataset) tornado;
-                ServiceUtil.deleteDataset(tDataset.getDatasetId(), this.username, this.userGroups);
-                if (ServiceUtil.deleteDataset(tDataset.getDatasetId(), this.username, this.userGroups) == null) {
-                    spaceRepository.addToOrphansSpace(tDataset.getDatasetId());
+                for (TornadoHazardDataset dataset: tDataset.getHazardDatasets()) {
+                    if (ServiceUtil.deleteDataset(dataset.getDatasetId(), this.username, this.userGroups) == null) {
+                        spaceRepository.addToOrphansSpace(dataset.getDatasetId());
+                    }
                 }
             }